// frontend/src/components/LoginPage/LoginForm/LoginForm.module.scss.d.ts
/*
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

declare const styles: {
  ackInlineLoading: string;
  loginCard: string;
  loginDesc: string;
  loginContainer: string;
  loginDescription: string;
  loginFormContent: string;
  loginFormInfo: string;
  queriusLogo: string;
  loginTitle: string;
  loginFormInputField: string;
  loginFormActions: string;
  signUpButton: string;
  loginNotification: string;
  apiKeyField: string;
  loader: string;
  loginFormIApiKeyField: string;
  infoIcon: string;
  tooltipContent: string;
};
export default styles;